benchmark/requirements.txt
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0 